restructure linter tests
ie split in multiple files
